add backtrack module
